Brings in 17 upstream commits, including the multi-DSL dispatcher
foundation (#90f807a) and the new CuTeDSL backend batches, which moved
several Triton launches out of ops/*.py into ops/backends/_triton/*.py.

Five files conflicted. All were resolved by taking upstream's version
wholesale and re-applying our device_context guards on top, so none of
the refactor is lost:

  cutedsl/ops/rms_norm.py, cutile/ops/swiglu.py,
  fused_linear_cross_entropy.py, fused_linear_jsd.py, jsd.py

jsd.py needs no guard any more -- its launch became a dispatch() call,
so the guard belongs in the backend impl instead.

Taking upstream's cutedsl/rms_norm.py reverted _cute_stream to its
zero-arg form while rms_norm_dw_reduce.py:126 still calls
_cute_stream(dW_partial.device); that would have been a TypeError at
runtime, so the device-aware signature is restored.

Guards added where launches now live:
  - backends/_triton/cross_entropy.py, backends/_triton/jsd.py (new
    dispatcher impls; cross_entropy's is an entry point, and jsd's is
    called unguarded from the CuTeDSL impl's fallback)
  - cutedsl/rms_norm.py: the five _launch_* helpers
  - cutedsl/swiglu.py, cutile/swiglu.py
  - fused_linear_cross_entropy.py, fused_linear_jsd.py

Verified: every upstream-changed Python file matches origin/main once
guard lines are filtered and whitespace ignored. The audit reports 6
remaining bare _cute_stream() calls (rope.py x4, the two sm90 scaled-CE
helpers); all sit in functions whose callers guard them.

+++ src/liger_kernel/ops/kl_div.py
@@ -69,8 +69,8 @@ def _kldiv_kernel_forward(
     for i in range(0, n_cols, BLOCK_SIZE):
         offsets = i + base_offsets
         mask = offsets < n_cols
-        y = tl.load(y_ptr + offsets, mask=mask, other=0.0)
-        y_true = tl.load(gt_ptr + offsets, mask=mask, other=0.0)
+        y = tl.load(y_ptr + offsets, mask=mask, other=0.0).to(tl.float32)
+        y_true = tl.load(gt_ptr + offsets, mask=mask, other=0.0).to(tl.float32)
 
         # KL(y_true || y) = y_true * (log(y_true) - log(y))
         # We compute KL(y_true || y) with y in the log-space
@@ -110,7 +110,7 @@ def _kldiv_kernel_backward(
         offsets = i + tl.arange(0, BLOCK_SIZE)
         mask = offsets < n_cols
 
-        target = tl.load(target_ptr + offsets, mask=mask, other=0.0)
+        target = tl.load(target_ptr + offsets, mask=mask, other=0.0).to(tl.float32)
 
         if not log_target:
             res = target * -1
